+++ ItemQualities/Assets/ItemQualities/Scripts/EntityStates/SprintArmorDash/SprintArmorDashIdle.cs
@@ -1,4 +1,5 @@
 using ItemQualities;
+using ItemQualities.ModCompatibility;
 using RoR2;
 
 namespace EntityStates.SprintArmorDash
@@ -10,6 +11,9 @@ namespace EntityStates.SprintArmorDash
         private float _lastValidInputTime = float.NegativeInfinity;
 
         private CharacterBody _attachedBody;
+        private new InputBankTest inputBank;
+
+        private bool canDash => _attachedBody && !_attachedBody.HasBuff(ItemQualitiesContent.Buffs.SprintArmorDashCooldown);
 
         public override void OnEnter()
         {
@@ -20,6 +24,7 @@ namespace EntityStates.SprintArmorDash
                 return;
 
             _attachedBody = networkedBodyAttachment.attachedBody;
+            inputBank = _attachedBody.inputBank;
         }
 
         public override void Update()
@@ -29,24 +34,57 @@ namespace EntityStates.SprintArmorDash
             if (!_attachedBody)
                 return;
 
-            if (isAuthority)
+            if (_attachedBody.hasEffectiveAuthority)
             {
                 UpdateAuthority();
             }
         }
 
+        public override void FixedUpdate()
+        {
+            base.FixedUpdate();
+
+            if (!_attachedBody)
+                return;
+
+            if (_attachedBody.hasEffectiveAuthority)
+            {
+                FixedUpdateAuthority();
+            }
+        }
+
         private void UpdateAuthority()
         {
-            if (!_attachedBody.HasBuff(ItemQualitiesContent.Buffs.SprintArmorDashCooldown) && _attachedBody.inputBank.rawMoveUp.justPressed)
+            if (canDash && inputBank.interact.down)
             {
-                float timeSinceLastInput = age - _lastValidInputTime;
-                if (timeSinceLastInput <= DoubleTapWindow)
+                if (inputBank.rawMoveUp.justPressed)
                 {
-                    outer.SetNextState(new SprintArmorDashDashingState());
-                }
+                    float timeSinceLastInput = age - _lastValidInputTime;
+                    if (timeSinceLastInput <= DoubleTapWindow)
+                    {
+                        StartDashAuthority();
+                    }
 
-                _lastValidInputTime = age;
+                    _lastValidInputTime = age;
+                }
             }
+            else
+            {
+                _lastValidInputTime = float.NegativeInfinity;
+            }
+        }
+
+        private void FixedUpdateAuthority()
+        {
+            if (canDash && RebindablesCompat.GetSprintArmorDashButtonState(inputBank).justPressed)
+            {
+                StartDashAuthority();
+            }
+        }
+
+        private void StartDashAuthority()
+        {
+            outer.SetNextState(new SprintArmorDashDashingState());
         }
     }
 }