+++ ItemQualities/Assets/ItemQualities/CHANGELOG.md
@@ -65,6 +65,11 @@
 * Infusion:
     * Fixed quality infusion health contributing to regular infusion's limit.
 
+* Rose Buckler:
+    * Default dash input changed to hold interact + double tap forward.
+    * Fixed dash cooldown being 1 second shorter than intended.
+    * Added support for a custom dash key if [Rebindables](https://thunderstore.io/c/riskofrain2/p/pseudopulse/Rebindables/) is installed.
+
 * Pocket I.C.B.M:
     * Now fires a small missile when using a skill that has at least 5s cooldown.